Audit report on Oracle Database Servers
We have completed an audit of the Oracle Database Servers for which fieldwork began March 10, 2023. Our audit was directed towards determining whether proper controls are in place within the database to prevent breaches of security, limit user access based on permissions, and logging/monitoring changes in the database. The audit was conducted in accordance with the Institute of Internal Auditors’ International Standards for the Professional Practice of Internal Auditing.
The scope of our audit included a complete review of the current controls over the availability and security of the Oracle database servers. Within this review, we focused on ensuring that the security controls in place mandated password parameter settings and password complexity. We also focused on making sure that there are proper controls in place to limit crucial privileges from being given to unauthorized users. Additionally, our audit scope included the investigation into the segregation of testing and production databases. We reviewed that a segregation exists, and validated if it is adequate, and that changes done in the testing database would require approval from privileged users before being launched into production database. Furthermore, the scope of our audit included a review of the controls in place to log and monitor data that is stored in the database. These tests included verifying whether or not the data is stored and encrypted properly, assuring that unnecessary data is not being collected, and that the data has a retention period specified in the records management. We also performed other audit procedures as deemed necessary to ascertain the degree of compliance with applicable laws, regulations, established policies and procedures, and to determine, where practical, if improvements and efficiency in operations could be gained.